requirements.txt
django==4.1.3
djangorestframework==3.14.0
pymysql==1.0.2
requests==2.28.2
